+++ platformio/commands/device/command.py
@@ -20,9 +20,9 @@ import click
 from serial.tools import miniterm
 
 from platformio import exception, fs, util
-from platformio.compat import dump_json_to_unicode, load_python_module
+from platformio.commands.device import helpers as device_helpers
+from platformio.compat import dump_json_to_unicode
 from platformio.managers.platform import PlatformFactory
-from platformio.project.config import ProjectConfig
 from platformio.project.exception import NotPlatformIOProjectError
 
 
@@ -135,7 +135,7 @@ def device_list(  # pylint: disable=too-many-branches
     help="Set the encoding for the serial port (e.g. hexlify, "
     "Latin1, UTF-8), default: UTF-8",
 )
-@click.option("--filter", "-f", multiple=True, help="Add text transformation")
+@click.option("--filter", "-f", multiple=True, help="Add filters/text transformations")
 @click.option(
     "--eol",
     default="CRLF",
@@ -174,15 +174,11 @@ def device_list(  # pylint: disable=too-many-branches
     help="Load configuration from `platformio.ini` and specified environment",
 )
 def device_monitor(**kwargs):  # pylint: disable=too-many-branches
-    click.echo(
-        "Looking for advanced Serial Monitor with UI? "
-        "Check http://bit.ly/pio-advanced-monitor"
-    )
     project_options = {}
     try:
         with fs.cd(kwargs["project_dir"]):
-            project_options = get_project_options(kwargs["environment"])
-        kwargs = apply_project_monitor_options(kwargs, project_options)
+            project_options = device_helpers.get_project_options(kwargs["environment"])
+        kwargs = device_helpers.apply_project_monitor_options(kwargs, project_options)
     except NotPlatformIOProjectError:
         pass
 
@@ -190,17 +186,17 @@ def device_monitor(**kwargs):  # pylint: disable=too-many-branches
     if "platform" in project_options:
         with fs.cd(kwargs["project_dir"]):
             platform = PlatformFactory.newPlatform(project_options["platform"])
-            register_platform_filters(platform, kwargs["project_dir"], kwargs["environment"])
+            device_helpers.register_platform_filters(
+                platform, kwargs["project_dir"], kwargs["environment"]
+            )
 
     if not kwargs["port"]:
         ports = util.get_serial_ports(filter_hwid=True)
         if len(ports) == 1:
             kwargs["port"] = ports[0]["port"]
         elif "platform" in project_options and "board" in project_options:
-            board_hwids = get_board_hwids(
-                kwargs["project_dir"],
-                platform,
-                project_options["board"],
+            board_hwids = device_helpers.get_board_hwids(
+                kwargs["project_dir"], platform, project_options["board"],
             )
             for item in ports:
                 for hwid in board_hwids:
@@ -217,12 +213,18 @@ def device_monitor(**kwargs):  # pylint: disable=too-many-branches
                 break
 
     # override system argv with patched options
-    sys.argv = ["monitor"] + options_to_argv(
+    sys.argv = ["monitor"] + device_helpers.options_to_argv(
         kwargs,
         project_options,
         ignore=("port", "baud", "rts", "dtr", "environment", "project_dir"),
     )
 
+    if not kwargs["quiet"]:
+        click.echo(
+            "Available filters and text transformations: %s"
+            % ", ".join(sorted(miniterm.TRANSFORMATIONS.keys()))
+        )
+        click.echo("More details at http://bit.ly/pio-monitor-filters")
     try:
         miniterm.main(
             default_port=kwargs["port"],
@@ -232,98 +234,3 @@ def device_monitor(**kwargs):  # pylint: disable=too-many-branches
         )
     except Exception as e:
         raise exception.MinitermException(e)
-
-def apply_project_monitor_options(cli_options, project_options):
-    for k in ("port", "speed", "rts", "dtr"):
-        k2 = "monitor_%s" % k
-        if k == "speed":
-            k = "baud"
-        if cli_options[k] is None and k2 in project_options:
-            cli_options[k] = project_options[k2]
-            if k != "port":
-                cli_options[k] = int(cli_options[k])
-    return cli_options
-
-
-def options_to_argv(cli_options, project_options, ignore=None):
-    result = project_options.get("monitor_flags", [])
-    for k, v in cli_options.items():
-        if v is None or (ignore and k in ignore):
-            continue
-        k = "--" + k.replace("_", "-")
-        if k in project_options.get("monitor_flags", []):
-            continue
-        if isinstance(v, bool):
-            if v:
-                result.append(k)
-        elif isinstance(v, tuple):
-            for i in v:
-                result.extend([k, i])
-        else:
-            result.extend([k, str(v)])
-    return result
-
-
-def get_project_options(environment=None):
-    config = ProjectConfig.get_instance()
-    config.validate(envs=[environment] if environment else None)
-    if not environment:
-        default_envs = config.default_envs()
-        if default_envs:
-            environment = default_envs[0]
-        else:
-            environment = config.envs()[0]
-    return config.items(env=environment, as_dict=True)
-
-
-def get_board_hwids(project_dir, platform, board):
-    with fs.cd(project_dir):
-        return (
-            platform
-            .board_config(board)
-            .get("build.hwids", [])
-        )
-
-class DeviceMonitorFilter(miniterm.Transform):
-    # NAME = "esp_exception_decoder" - all filters must have one
-
-    # Called by PlatformIO to pass context
-    def __init__(self, project_dir, environment):
-        super(DeviceMonitorFilter, self).__init__()
-
-        self.project_dir = project_dir
-        self.environment = environment
-
-        self.config = ProjectConfig.get_instance()
-        if not self.environment:
-            default_envs = self.config.default_envs()
-            if default_envs:
-                self.environment = default_envs[0]
-            else:
-                self.environment = self.config.envs()[0]
-
-    # Called by the miniterm library when the filter is acutally used
-    def __call__(self):
-        return self
-
-def register_platform_filters(platform, project_dir, environment):
-    monitor_dir = os.path.join(platform.get_dir(), "monitor")
-    if not os.path.isdir(monitor_dir):
-        return
-
-    for fn in os.listdir(monitor_dir):
-        if not fn.startswith("filter_") or not fn.endswith(".py"):
-            continue
-        path = os.path.join(monitor_dir, fn)
-        if not os.path.isfile(path):
-            continue
-
-        dot = fn.find(".")
-        module = load_python_module("platformio.commands.device.%s" % fn[:dot], path)
-        for key in dir(module):
-            member = getattr(module, key)
-            try:
-                if issubclass(member, DeviceMonitorFilter) and hasattr(member, "NAME"):
-                    miniterm.TRANSFORMATIONS[member.NAME] = member(project_dir, environment)
-            except TypeError:
-                pass